+++ src/app/wallet/create/page.tsx
@@ -1,0 +1,10 @@
+import CreateWalletPage from "@/components/pages/createWallet";
+
+const WalletPage = () => {
+  // document.getElementById('delete_wallet_modal').showModal()
+  return (
+    <CreateWalletPage />
+  );
+};
+
+export default WalletPage;
