+++ src/app/wallet/delete/page.tsx
@@ -1,0 +1,10 @@
+import DeleteWalletPage from "@/components/pages/deleteWallet";
+
+const WalletPage = () => {
+  // document.getElementById('delete_wallet_modal').showModal()
+  return (
+    <DeleteWalletPage />
+  );
+};
+
+export default WalletPage;
